# openapi.yaml
openapi: 3.0.0
info:
  description: |
    This is a MyTaskApp api documentation BE20 altera academy.  You can find
    out more about Swagger at
    [http://swagger.io](http://swagger.io) or on
    [irc.freenode.net, #swagger](http://swagger.io/irc/).
  version: "1.0.0"
  title: MyTaskApp
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/DWIMIRANDAA/MyTaskApp/1.0.0
  - url: 'https://petstore.swagger.io/v2'
...
paths:
  /users:
    post:
      tags:
        - Users
      summary: Register
      responses:
        default:
          description: successful operation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      description: Operasi untuk registrasi pengguna.
  

    get:
      tags:
        - Users
      summary: Get User Profile
      description: Mendapatkan profil pengguna yang sedang login.
      parameters:
        - name: name
          in: path
          description: The name that needs to be fetched. Use user1 for testing.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found

    put:
      tags:
        - Users
      summary: Update User Profile
      description: Mengupdate data pengguna berdasarkan data token ID yang login.
      parameters:
        - name: name
          in: path
          description: name that need to be updated
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Update successful
        '400':
          description: Invalid user supplied
        '404':
          description: User not found
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: Updated user object
        required: true

    delete:
      tags:
        - Users
      summary: Delete User Account
      description: Menghapus data pengguna berdasarkan data token ID yang login.
      operationId: deleteUser
      parameters:
        - name: name
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Invalid name 
        '404':
          description: User not found
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone number: 
          type: string
        addrres: 
          type: string
        role : 
          type: string
        password:
          type: string
      required: 
        - name
        - email
        - phone number
        - address
        - role
        - password
...